% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tell_me_how_drunk.R
\name{check_age}
\alias{check_age}
\title{Check legal drinking age in Germany}
\usage{
check_age(age, drinks)
}
\arguments{
\item{age}{numerical vector indicating the ages of the drinkers}

\item{drinks}{list of drinks}
}
\value{
no return, function issues a warning in case of illegal drinking
}
\description{
Check legal drinking age in Germany
}
