% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tell_me_how_drunk.R
\name{check_drinks_input}
\alias{check_drinks_input}
\title{Input sanitation for drinks}
\usage{
check_drinks_input(drinks)
}
\arguments{
\item{drinks}{named integer vector containing drinks}
}
\value{
named integer vector of valid entries in drinks
}
\description{
Input sanitation for drinks
}
