% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promillo-doc.R
\name{check_illegal_drink}
\alias{check_illegal_drink}
\title{check legality of drinking age}
\usage{
check_illegal_drink(age, drinks)
}
\arguments{
\item{age}{age in years}

\item{drinks}{possible drinks}
}
\value{
total mass of alcohol consumed
}
\description{
check legality of drinking age
}
