% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promillo.R
\name{get_alcohol}
\alias{get_alcohol}
\title{Calculate the alcohol level of some drinks}
\usage{
get_alcohol(drinks)
}
\arguments{
\item{drinks}{list or vector of drinks, named numeric, supported drinks are
"massn", "hoibe", "wein", "schnaps"}
}
\value{
blood alcohol level, numeric
}
\description{
Based on the drinks supplied, calculate the overall alcohol level.
}
