% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tell_me_how_drunk.R
\name{get_bodywater}
\alias{get_bodywater}
\title{calculate the bodywater using basic body properties}
\usage{
get_bodywater(sex = c("male", "female"), age, height, weight)
}
\arguments{
\item{sex}{single character value, either "male" or "female"}

\item{age}{single numeric value}

\item{height}{single numeric value}

\item{weight}{single numeric value}
}
\value{
single numeric value
}
\description{
Whatson's formula is used with Eickner's modification to include the
age dependency for females.
}
