% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tell_me_how_drunk.R
\name{hard_drinks}
\alias{hard_drinks}
\title{Decide, whether hard drinks are present in a list of drinks}
\usage{
hard_drinks(drinks)
}
\arguments{
\item{drinks}{list of drinks}
}
\value{
boolean indicating whether hard drinks are present
}
\description{
Decide, whether hard drinks are present in a list of drinks
}
