% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promillo.R
\name{show_me_how_drunk}
\alias{show_me_how_drunk}
\title{Plot the permille over the drinking time}
\usage{
show_me_how_drunk(age, sex, height, weight, drinking_time, drinks)
}
\arguments{
\item{age}{age in years, numeric}

\item{sex}{sex ("male" vs. "female"), character}

\item{height}{height in cm, numeric}

\item{weight}{weight in kg, numeric}

\item{drinking_time}{vector of start and end time, POSIXct}

\item{drinks}{list or vector of drinks, named numeric, supported drinks are
"massn", "hoibe", "wein", "schnaps"}
}
\value{
qplot plotting the permille over the drinking time (x-axis in five
minutes intervals)
}
\description{
Plot the permille over the drinking time with the x-axis resembling
intervals of five minutes.
}
