% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tell_me_how_drunk.R
\name{tell_me_alcohol_amount}
\alias{tell_me_alcohol_amount}
\title{Compute amount of alcohol contained in drinks}
\usage{
tell_me_alcohol_amount(drinks)
}
\arguments{
\item{drinks}{named integer vector containing drinks}
}
\value{
double denoting the amount of alcohol in grams contained in drinks
}
\description{
Compute amount of alcohol contained in drinks
}
