% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tell_me_how_drunk.R
\name{tell_me_amount_body_water}
\alias{tell_me_amount_body_water}
\title{Compute amount of body water in a human}
\usage{
tell_me_amount_body_water(age, height, weight, sex = c("male", "female"))
}
\arguments{
\item{age}{numeric type denoting the age}

\item{height}{numeric type denoting the height in cm}

\item{weight}{numeric type denoting the weight in kg}

\item{sex}{sex of person}
}
\description{
Compute amount of body water in a human
}
