% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tell_me_how_drunk.R
\name{tell_me_blood_alcohol_level}
\alias{tell_me_blood_alcohol_level}
\title{Computes blood alcohol level for a person}
\usage{
tell_me_blood_alcohol_level(alcohol_amount, weight, distribution_factor)
}
\arguments{
\item{alcohol_amount}{amount of alcohol consumed}

\item{weight}{weight in kg}

\item{distribution_factor}{distribution factor}
}
\description{
Computes blood alcohol level for a person
}
