% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tell_me_how_drunk.R
\name{tell_me_current_blood_alcohol_level}
\alias{tell_me_current_blood_alcohol_level}
\title{Computes the current blood alcohol level based on drinking time}
\usage{
tell_me_current_blood_alcohol_level(blood_alcohol_level, drinking_time)
}
\arguments{
\item{blood_alcohol_level}{blood alcohol level at time of drinking}

\item{drinking_time}{POSIXct timestamps containing start and end of drinking}
}
\value{
blood alcohol level at the end of drinking
}
\description{
Computes the current blood alcohol level based on drinking time
}
