% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tell_me_how_drunk.R
\name{tell_me_distribution_factor}
\alias{tell_me_distribution_factor}
\title{Computes distribution factor}
\usage{
tell_me_distribution_factor(amount_body_water, weight)
}
\arguments{
\item{amount_body_water}{Amount of body water in kg}

\item{weight}{weight in kg}
}
\value{
distribution factor
}
\description{
Computes distribution factor
}
