% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promillo.R
\name{tell_me_how_drunk}
\alias{tell_me_how_drunk}
\title{Calculate the blood alcohol level}
\usage{
tell_me_how_drunk(age, sex = c("male", "female"), height, weight,
  drinking_time, drinks)
}
\arguments{
\item{age}{age in years, numeric}

\item{sex}{sex ("male" vs. "female"), character}

\item{height}{height in cm, numeric}

\item{weight}{weight in kg, numeric}

\item{drinking_time}{vector of start and end time, POSIXct}

\item{drinks}{list or vector of drinks, named numeric, supported drinks are
"massn", "hoibe", "wein", "schnaps"}
}
\value{
blood alcohol level, numeric
}
\description{
Calculate the blood alcohol level depending on your age, sex, height,
weight, drinking_time, and the drinks you had.
}
