% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promillo.R
\name{get_alcohol}
\alias{get_alcohol}
\title{Calculates the absorbed alcohol mass}
\usage{
get_alcohol(drinks)
}
\arguments{
\item{drinks}{List containing the drinks. Each list element must contain two elements namely the
name of the drink and the quantity consumend. The drinks must be a subset of ("massn", "hoibe", "wein", "schnaps")}
}
\value{
Double containing the absorbed alcohol mass
}
\description{
Calculates the absorbed alcohol mass
}
