% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promillo.R
\name{get_alcohol}
\alias{get_alcohol}
\title{Compute total amount of alcohol intake.}
\usage{
get_alcohol(drinks)
}
\arguments{
\item{drinks}{character vector of predefined drinks}
}
\description{
Compute total amount of alcohol intake.
}
