% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alcohol.R
\name{get_alcohol}
\alias{get_alcohol}
\title{Compute the amount of consumed alcohol}
\usage{
get_alcohol(drinks)
}
\arguments{
\item{drinks}{list or vector with names "massn", "hoibe", "wein", "schnaps"
counting the number consumed of each type of drink}
}
\value{
single numeric value of consumed alcohol in ml
}
\description{
Calculates the amount of consumed alcohol based on 4 alcohol drinks: "massn",
"hoibe", "wein", "schnaps"
}
