% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promillo-doc.R
\name{get_alcohol}
\alias{get_alcohol}
\title{compute total mass of alcohol consumed}
\usage{
get_alcohol(drinks)
}
\arguments{
\item{drinks}{possible drinks}
}
\value{
total mass of alcohol consumed
}
\description{
compute total mass of alcohol consumed
}
