% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tell_me_how_drunk.R
\name{get_alcohol}
\alias{get_alcohol}
\title{compute consumed alcohol in g}
\usage{
get_alcohol(drinks)
}
\arguments{
\item{drinks}{list or vector with names "massn", "hoibe", "wein", "schnaps"}
}
\value{
consumed alcohol in g
}
\description{
compute consumed alcohol in g
}
