% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tell_me_how_drunk.R
\name{get_alcohol}
\alias{get_alcohol}
\title{calculate the amount of alcohol for a set of alcohols drunk}
\usage{
get_alcohol(drinks)
}
\arguments{
\item{drinks}{list or vector named with possible drink choices "massn", "hoibe", "wein", and "schnaps" and the number of portions drunk}
}
\value{
single numeric value
}
\description{
For each of the possible types of drink, a fixed volume of one portion and
alcohol concentration is assumed. The alcohol density is fixed at 0.8.
}
