% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promillo.R
\name{get_alcohol}
\alias{get_alcohol}
\title{get alcohol consumption}
\usage{
get_alcohol(drinks)
}
\arguments{
\item{drinks}{drinks vector, e.g., c("schnaps", "wein")}
}
\value{
total alcohol consumed
}
\description{
get alcohol consumption
}
