% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promillo-nodoc.R
\name{get_alcohol}
\alias{get_alcohol}
\title{Alcohol mass}
\usage{
get_alcohol(drinks)
}
\arguments{
\item{drinks}{a named numeric vector of drinks. Possible choices
"massn", "hoibe", "wein" and "schnaps"}
}
\value{
The function returns a numeric value indicating the alcohol mass.
}
\description{
The function calculates how much alcohol a person consumed.
}
