% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promillo.R
\name{get_bodywater}
\alias{get_bodywater}
\title{Compute amount of bodywater for a person.}
\usage{
get_bodywater(sex = c("male", "female"), age, height, weight)
}
\arguments{
\item{sex}{male or female}

\item{age}{positive integer between 10 and 110}

\item{height}{in cm}

\item{weight}{in kg}
}
\description{
Compute amount of bodywater for a person.
}
