% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bodywater.R
\name{get_bodywater}
\alias{get_bodywater}
\title{Compute body water index}
\usage{
get_bodywater(sex = c("male", "female"), age, height, weight)
}
\arguments{
\item{sex}{single character as "male" or "female"}

\item{age}{age of a person in years}

\item{height}{height of a person in cm}

\item{weight}{weight of a person in kg}
}
\value{
single numeric value of bodywater index
}
\description{
Calculates the body water index based on the physical properties of a person
}
