% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promillo.R
\name{get_bodywater}
\alias{get_bodywater}
\title{Calculate the bodywater level}
\usage{
get_bodywater(sex = c("male", "female"), age, height, weight)
}
\arguments{
\item{sex}{sex ("male" vs. "female"), character}

\item{age}{age in years, numeric}

\item{height}{height in cm, numeric}

\item{weight}{weight in kg, numeric}
}
\value{
bodywater level, numeric
}
\description{
Based on sex, age, height and weight, calculate the bodywater level.
}
