% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promillo-doc.R
\name{get_bodywater}
\alias{get_bodywater}
\title{compute total amount of water in the body}
\usage{
get_bodywater(sex = c("male", "female"), age, height, weight)
}
\arguments{
\item{sex}{sex as a character ("male" or "female")}

\item{age}{age in years}

\item{height}{height in cm}

\item{weight}{weight in kg}
}
\value{
total amount of water in the body
}
\description{
compute total amount of water in the body
}
