% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tell_me_how_drunk.R
\name{get_bodywater}
\alias{get_bodywater}
\title{compute amount of water in a human body}
\usage{
get_bodywater(sex = c("male", "female"), age, height, weight)
}
\arguments{
\item{sex}{male or female (character)}

\item{age}{in years}

\item{height}{in cm}

\item{weight}{in kg}
}
\value{
amount of water in a human body
}
\description{
compute amount of water in a human body
}
