% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promillo.R
\name{get_bodywater}
\alias{get_bodywater}
\title{get bodywater of a person}
\usage{
get_bodywater(sex = c("male", "female"), age, height, weight)
}
\arguments{
\item{sex}{person's sex (male or female)}

\item{age}{person's age in years}

\item{height}{person's height in cm}

\item{weight}{person's weigh in Kg}
}
\value{
bodywater
}
\description{
get bodywater of a person
}
