% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promillo-nodoc.R
\name{get_bodywater}
\alias{get_bodywater}
\title{Body Water of a person}
\usage{
get_bodywater(sex = c("male", "female"), age, height, weight, drinks)
}
\arguments{
\item{sex}{single character. Possible choices are "male" and "female".}

\item{age}{the age of the person. Single number.}

\item{height}{the height in cm of that person. Single number.}

\item{weight}{the weight in kg of that person. Single number.}

\item{drinks}{a named numeric vector of drinks. Possible choices
"massn", "hoibe", "wein" and "schnaps"}
}
\value{
Function returns the calculated Body Water.
}
\description{
The function calculates the body water of a specific person using age,
sex, height, weight. In addition the function performs an age check using
the age of a person and the drinks said person consumed.
}
\details{
\code{get_bodywater} uses the Whatson formula. see references.
}
\references{
\href{https://web.archive.org/web/20150123143123/http://promille-rechner.org/erlaeuterung-der-promille-berechnung/}{Erlaeuterung der Promille-Berechnung}
}
