% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promillo.R
\name{get_permille}
\alias{get_permille}
\title{Calculates the alcohol level}
\usage{
get_permille(alcohol_drunk, bodywater, drinking_time)
}
\arguments{
\item{alcohol_drunk}{Absorbed alcohol mass. Calculated in \code{\link{get_alcohol}}}

\item{bodywater}{Estimated bodywater of the person. Calculated in \code{\link{get_bodywater}}}

\item{drinking_time}{Vector of length 2 containing two posixct indicating start and end of drinking}
}
\value{
Numeric value for blood alocohol level [in per mille]
}
\description{
Calculates the alcohol level
}
