% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promillo.R
\name{get_permille}
\alias{get_permille}
\title{Compute blood alcohol content in permille.}
\usage{
get_permille(alcohol_drunk, bodywater, drinking_time)
}
\arguments{
\item{alcohol_drunk}{total amount of alcohol intake}

\item{bodywater}{total amount of bodywater}

\item{drinking_time}{vector of start and end time of drinking}
}
\description{
Compute blood alcohol content in permille.
}
