% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_permille.R
\name{get_permille}
\alias{get_permille}
\title{Compute permille value}
\usage{
get_permille(alcohol_drunk, bodywater, drinking_time)
}
\arguments{
\item{alcohol_drunk}{single numeric value of consumed alcohol}

\item{bodywater}{single numeric value of the body water index of a person}

\item{drinking_time}{sorted POSIXct vector giving start and end of the party}
}
\value{
single numeric value of permille
}
\description{
Calculates the BAC
}
