% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promillo.R
\name{get_permille}
\alias{get_permille}
\title{Calculate the permille}
\usage{
get_permille(alcohol_drunk, bodywater, drinking_time)
}
\arguments{
\item{alcohol_drunk}{alcohol level, see \link{get_alcohol}}

\item{bodywater}{bodywater level, see \link{get_bodywater}}

\item{drinking_time}{vector of start and end time, POSIXct}
}
\description{
Using the alcohol level and bodywater level, calculate the permille.
}
