% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promillo-doc.R
\name{get_permille}
\alias{get_permille}
\title{compute BAC per-mille}
\usage{
get_permille(alcohol_drunk, bodywater, drinking_time)
}
\arguments{
\item{alcohol_drunk}{total mass of alcohol consumed}

\item{bodywater}{total amount of water in the body}

\item{drinking_time}{duration of drinking time in hours}
}
\value{
blood alcohol concentration (BAC) per-mille
}
\description{
compute BAC per-mille
}
