% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tell_me_how_drunk.R
\name{get_permille}
\alias{get_permille}
\title{compute BAC at \code{drinking_time[2]}}
\usage{
get_permille(alcohol_drunk, bodywater, drinking_time)
}
\arguments{
\item{alcohol_drunk}{consumed alcohol in g}

\item{bodywater}{amount of water in a human body}

\item{drinking_time}{sorted POSIXct vector giving start and end of the party}
}
\value{
BAC at \code{drinking_time[2]}
}
\description{
compute BAC at \code{drinking_time[2]}
}
