% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promillo.R
\name{get_permille}
\alias{get_permille}
\title{get Permille}
\usage{
get_permille(alcohol_drunk, bodywater, drinking_time)
}
\arguments{
\item{alcohol_drunk}{amount of alcohol consumed}

\item{bodywater}{person's bodywater}

\item{drinking_time}{interval of time during which alcohol was consumed}
}
\value{
permille blood alcohol value
}
\description{
get Permille
}
