% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_me_how_drunk.R
\name{show_me_how_drunk}
\alias{show_me_how_drunk}
\title{Plotting alcohol value during drinking time}
\usage{
show_me_how_drunk(age, sex, height, weight, drinking_time, drinks)
}
\arguments{
\item{age}{Number between 10 and 110.}

\item{sex}{Character that matches either "male" or "female".}

\item{height}{Number between 100 and 230 [in cm]}

\item{weight}{Number between 40 and 300 [in kg]}

\item{drinking_time}{Vector of length 2 containing two posixct indicating start and end of drinking}

\item{drinks}{List containing the drinks. Each list element must contain two elements namely the
name of the drink and the quantity consumend. The drinks must be a subset of ("massn", "hoibe", "wein", "schnaps")}
}
\value{
qplot visualizing the alcohol value [in per mille] during drinking time
}
\description{
Plots the calculated alcohol level at 5min intervals over the entire drinking time.
}
\seealso{
\code{\link{tell_me_how_drunk}} for the underlying calulation of alcohol level
}
