% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promillo-doc.R
\name{show_me_how_drunk}
\alias{show_me_how_drunk}
\title{Show me how drunk}
\usage{
show_me_how_drunk(
  age,
  sex = c("male", "female"),
  height,
  weight,
  drinking_time,
  drinks
)
}
\arguments{
\item{age}{age in years}

\item{sex}{sex as a character ("m", "w", "male", "female"; lower/upper case both
acceptable)}

\item{height}{height in cm}

\item{weight}{weight in kg}

\item{drinking_time}{start and end time points of alcohol consumption as a
POSIXct vector (sorted ascended)}

\item{drinks}{vector or list with number of consumed drinks
("massn", "hoibe", "wein", "schnaps" allowed)}
}
\description{
Plots per-mille blood alcohol concentration (BAC) computed using \code{\link[=tell_me_how_drunk]{tell_me_how_drunk()}}
at 5 mins interval
}
