% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_me_how_drunk.R
\name{show_me_how_drunk}
\alias{show_me_how_drunk}
\title{plot promille versus 5 minute time stamps of drinking time}
\usage{
show_me_how_drunk(age, sex, height, weight, drinking_time, drinks)
}
\arguments{
\item{age}{in years}

\item{sex}{male or female (character)}

\item{height}{in cm}

\item{weight}{in kg}

\item{drinking_time}{sorted POSIXct vector giving start and end of the party}

\item{drinks}{list or vector with names "massn", "hoibe", "wein", "schnaps"}
}
\value{
ggplot2 object
}
\description{
create line plot of promille versus 5 minute time stamps of drinking time
}
\examples{
show_me_how_drunk(
  age = 39,
  sex = "male",
  height = 190,
  weight = 85,
  drinking_time = c(
    as.POSIXct("2021-01-28 14:00:00"),
    as.POSIXct("2021-01-28 16:00:00")
  ),
  drinks = c(schnaps = 1, hoibe = 3)
)
}
