% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_me_how_drunk.R
\name{show_me_how_drunk}
\alias{show_me_how_drunk}
\title{hypothesize course of blood alcohol during a party}
\usage{
show_me_how_drunk(age, sex, height, weight, drinking_time, drinks)
}
\arguments{
\item{age}{single numeric value}

\item{sex}{single character value, either "male" or "female"}

\item{height}{single numeric value}

\item{weight}{single numeric value}

\item{drinking_time}{vector of two POSIXct values indicating when person started and ended drinking}

\item{drinks}{list or vector named with "massn", "hoibe", "wein", or "schnaps" containing the number of portions drunk}
}
\value{
ggplot object
}
\description{
For simplicity, it is assumed that the drinking speed is constant and
the alcohol amount calculated is based on a mix of all drinks drunk during
the party.
}
