% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promillo.R
\name{show_me_how_drunk}
\alias{show_me_how_drunk}
\title{show me how drunk, plots per mille alcohol level at 5 mins interval}
\usage{
show_me_how_drunk(age, sex, height, weight, drinking_time, drinks)
}
\arguments{
\item{age}{person's in years}

\item{sex}{person's sex (male or female)}

\item{height}{person's height in cm}

\item{weight}{person's weight in Kg}

\item{drinking_time}{drinking time in hours}

\item{drinks}{drinks vector e.g., c("schnaps", "wein")}
}
\description{
show me how drunk, plots per mille alcohol level at 5 mins interval
}
