% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_me_how_drunk.R
\name{show_me_how_drunk}
\alias{show_me_how_drunk}
\title{Plotfunction for permille}
\usage{
show_me_how_drunk(age, sex, height, weight, drinking_time, drinks)
}
\arguments{
\item{age}{the age of the person. Single number.}

\item{sex}{single character. Possible choices are "male" and "female".}

\item{height}{the height in cm of that person. Single number.}

\item{weight}{the weight in kg of that person. Single number.}

\item{drinking_time}{a POSIXct vector of length 2. First argument is the
time the person started to drink alcohol, second argument is the end of
drinking.}

\item{drinks}{a named numeric vector of drinks. Possible choices
"massn", "hoibe", "wein" and "schnaps"}
}
\value{
qplot
}
\description{
The function plots the decline of blood-level alcohol over time in 5-minute steps.
}
