% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tell_me_how_drunk.R
\name{tell_me_how_drunk}
\alias{tell_me_how_drunk}
\title{Computes blood alcohol level for a person}
\usage{
tell_me_how_drunk(age, sex = c("male", "female"), height, weight,
  drinking_time, drinks)
}
\arguments{
\item{age}{Age of person (numeric)}

\item{sex}{Sex of person ("male", "female")}

\item{height}{Height of person in cm}

\item{weight}{Weight of person in kg}

\item{drinking_time}{POSIXct timestamps containing start and end of drinking}

\item{drinks}{named integer vector containing drinks consumed}
}
\value{
blood alcohol level at the end of drinking
}
\description{
Computes blood alcohol level for a person
}
