% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promillo.R
\name{tell_me_how_drunk}
\alias{tell_me_how_drunk}
\title{Calculates blood alcohol level.}
\usage{
tell_me_how_drunk(age, sex = c("male", "female"), height, weight,
  drinking_time, drinks)
}
\arguments{
\item{age}{Number between 10 and 110.}

\item{sex}{Character that matches either "male" or "female".}

\item{height}{Number between 100 and 230 [in cm]}

\item{weight}{Number between 40 and 300 [in kg]}

\item{drinking_time}{Vector of length 2 containing two posixct indicating start and end of drinking}

\item{drinks}{List containing the drinks. Each list element must contain two elements namely the
name of the drink and the quantity consumend. The drinks must be a subset of ("massn", "hoibe", "wein", "schnaps")}
}
\value{
Numeric value for blood alocohol level [in per mille]
}
\description{
Calculates the current blood alcohol level based on a variety of drinks.
Formula ist based on \url{https://web.archive.org/web/20150123143123/http://promille-rechner.org/erlaeuterung-der-promille-berechnung/}
}
\examples{
tell_me_how_drunk(age = 39, sex = "male", height = 190, weight = 87,
drinking_time = as.POSIXct(c("2016-10-03 17:15:00", "2016-10-03 22:55:00")),
drinks = c("massn" = 3, "schnaps" = 4))

}
