% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promillo.R
\name{tell_me_how_drunk}
\alias{tell_me_how_drunk}
\title{Compute blood alcohol content in permille.}
\usage{
tell_me_how_drunk(age, sex = c("male", "female"), height, weight,
  drinking_time, drinks)
}
\arguments{
\item{age}{positive integer between 10 and 110}

\item{sex}{male or female}

\item{height}{in cm}

\item{weight}{in kg}

\item{drinking_time}{vector of start and end time of drinking}

\item{drinks}{character vector of predefined drinks}
}
\value{
blood alcohol content in permille
}
\description{
Compute blood alcohol content in permille.
}
\examples{
drinking_time <- as.POSIXct(c("2016-10-03 17:15:00", "2016-10-03 22:55:00"))
drinks <- c("wein" = 2)
tell_me_how_drunk(21, "female",
height = 172,
weight = 57,
drinking_time = drinking_time,
drinks = drinks)
}
