% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tell_me_how_drunk.R
\name{tell_me_how_drunk}
\alias{tell_me_how_drunk}
\title{Computation of Body Alcohol Concentration (BAC)}
\usage{
tell_me_how_drunk(age, sex = c("male", "female"), height, weight,
  drinking_time, drinks)
}
\arguments{
\item{age}{age of a person in years}

\item{sex}{single character "male" or "female"}

\item{height}{height of a person in cm}

\item{weight}{weight of a person in kg}

\item{drinking_time}{sorted POSIXct vector giving start and end of the party}

\item{drinks}{list or vector with names like "massn", "hoibe", "wein", "schnaps"
counting the consumed number of each type of beverage}
}
\value{
approximate BAC in per mille
}
\description{
Computes approximate BAC (in per mille) at the end of the party
as described \href{https://web.archive.org/web/20150123143123/http://promille-rechner.org/erlaeuterung-der-promille-berechnung/}{here}.
}
