% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promillo-doc.R
\name{tell_me_how_drunk}
\alias{tell_me_how_drunk}
\title{Tell me how drunk}
\usage{
tell_me_how_drunk(
  age,
  sex = c("male", "female"),
  height,
  weight,
  drinking_time,
  drinks
)
}
\arguments{
\item{age}{age in years}

\item{sex}{sex as a character ("m", "w", "male", "female"; lower/upper case both
acceptable)}

\item{height}{height in cm}

\item{weight}{weight in kg}

\item{drinking_time}{start and end time points of alcohol consumption as a
POSIXct vector (sorted ascended)}

\item{drinks}{vector or list with number of consumed drinks
("massn", "hoibe", "wein", "schnaps" allowed)}
}
\value{
blood alcohol concentration in per-mille, adjusted for the reduction
of the blood alcohol level at the end of drinking time
}
\description{
Computes the blood alcohol concentration (BAC) using the Widmark and Whatson
combination
}
\references{
https://web.archive.org/web/20150123143123/http://promille-rechner.org/erlaeuterung-der-promille-berechnung/
}
\author{
Anastasiia Holovchak
}
