% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tell_me_how_drunk.R
\name{tell_me_how_drunk}
\alias{tell_me_how_drunk}
\title{calculate the blood alcohol per mille at the end of a party}
\usage{
tell_me_how_drunk(
  age,
  sex = c("male", "female"),
  height,
  weight,
  drinking_time,
  drinks
)
}
\arguments{
\item{age}{single numeric value}

\item{sex}{single character value, either "male" or "female"}

\item{height}{single numeric value}

\item{weight}{single numeric value}

\item{drinking_time}{vector of two POSIXct values indicating when person started and ended drinking}

\item{drinks}{list or vector named with possible drink choices "massn", "hoibe", "wein", and "schnaps" and the number of portions drunk}
}
\value{
blood alcohol per mille as single numeric value
}
\description{
The calculation is based on Whatson's and Widmark's formula.
}
