% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promillo.R
\name{tell_me_how_drunk}
\alias{tell_me_how_drunk}
\title{Tell me how drunk}
\usage{
tell_me_how_drunk(
  age,
  sex = c("male", "female"),
  height,
  weight,
  drinking_time,
  drinks
)
}
\arguments{
\item{age}{person's in years}

\item{sex}{person's sex (male or female)}

\item{height}{person's height in cm}

\item{weight}{person's weight in Kg}

\item{drinking_time}{drinking time in hours}

\item{drinks}{drinks vector e.g., c("schnaps", "wein")}
}
\value{
promille Per mille blood alcohol value
}
\description{
Tell me how drunk
}
