% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promillo-nodoc.R
\name{tell_me_how_drunk}
\alias{tell_me_how_drunk}
\title{Bloodlevel alcohol of a person in permille}
\usage{
tell_me_how_drunk(
  age,
  sex = c("male", "female"),
  height,
  weight,
  drinking_time,
  drinks
)
}
\arguments{
\item{age}{the age of the person. Single number.}

\item{sex}{single character. Possible choices are "male" and "female".}

\item{height}{the height in cm of that person. Single number.}

\item{weight}{the weight in kg of that person. Single number.}

\item{drinking_time}{a POSIXct vector of length 2. First argument is the
time the person started to drink alcohol, second argument is the end of
drinking.}

\item{drinks}{a named numeric vector of drinks. Possible choices
"massn", "hoibe", "wein" and "schnaps"}
}
\value{
The function returns the permille value after ending the drinking
process
}
\description{
This function calculates the blood level alcohol in permille using height,
weight, sex, the time of drinking, age an number of specific drinks.
}
\examples{
tell_me_how_drunk(
  age = 39, sex = "male", height = 190, weight = 87,
  drinking_time = as.POSIXct(c("2016-10-03 17:15:00", "2016-10-03 22:55:00")),
  drinks = c("massn" = 3, "schnaps" = 4)
)
}
